projects/aa-semennikov/lab_1/requirements.txt
annotated-types-0.7.0
anyio-4.11.0
distro-1.9.0
h11-0.16.0
httpcore-1.0.9
httpx-0.28.1
jiter-0.11.1
openai-2.6.1
pydantic-2.12.3
pydantic-core-2.41.4
sniffio-1.3.1
typing-extensions-4.15.0
typing-inspection-0.4.2